+++ ouroboros-network/CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Non-breaking changes
 
 * Fix random selection of peers to peershare with.
+* Fixed bug where peers with 'DoNotAdvertisePeer' flag were being shared
+* Fixed peer sharing pool of peers to be shared being confused with the pool
+  of peers to request to.
 
 ## 0.10.0.1 -- 2023-11-16